requirements.txt
streamlit
anthropic
python-dotenv
streamlit-code-editor